+++ src/rent/selector/SelectorRent.js
@@ -65,6 +65,10 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
     this.family_id = 'family_id';
     this.family_id_selector = '#family_id';
     this.family_selector = '.family';   
+    // Rental Location Code
+    this.rental_location_code = 'rental_location_code',
+    this.rental_location_code_selector = '#rental_location_code',   
+    this.rental_location_selector = '.rental_location', 
     // Accept age
     this.accept_age_selector = '#accept_age';
 
@@ -82,6 +86,7 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
     this.pickupHours = []; // Available pickup hours
     this.returnHours = []; // Available return hours
     this.families = []; // Families
+    this.rentalLocations = []; // Rental Locations
 
     this.dateToMinDate = null;
 
@@ -116,6 +121,36 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
     };
 
     /**
+     * Load rental locations
+     */ 
+    this.loadRentalLocations = function() { /* Load rental locations */
+
+      console.log('loadRentalLocations');
+      var self = this;
+      var url = commonServices.URL_PREFIX + '/api/booking/frontend/rental-locations';
+      if (commonServices.apiKey && commonServices.apiKey != '') {
+        url += '?api_key='+commonServices.apiKey;
+      }  
+      // Request
+      $.ajax({
+        type: 'GET',
+        url: url,
+        dataType: 'json',
+        success: function(data, textStatus, jqXHR) {
+          self.rentalLocations = data;
+          for (var idx=0;idx<self.rentalLocations.length;idx++){
+            self.rentalLocations[idx]['id'] = self.rentalLocations[idx]['code'];
+            self.rentalLocations[idx]['description'] =  self.rentalLocations[idx]['name'];
+          }          
+          self.selectorView.update('rentalLocations', null);
+        },
+        error: function(data, textStatus, jqXHR) {
+          alert(i18next.t('selector.error_loading_data'));
+        }
+      });      
+    };    
+
+    /**
      * Access the API to get the available pickup days in a month
      */
     this.loadPickupDays = function(year, month) { /* Load pickup days */
@@ -131,6 +166,12 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
       if (this.configuration.pickupReturnPlace && $(this.pickup_place_selector).val() != '') {
         url += '&place='+$(this.pickup_place_selector).val();
       }
+      if (this.selectorView.applyRentalLocationSelector() && $(this.rental_location_code_selector).val() != '') {
+        url += '&rental_location_code='+$(this.rental_location_code_selector).val();
+      }  
+      else if (this.selectorView.usedFixedRentalLocation() && this.selectorView.fixedRentalLocationValue() != '') {
+        url += '&rental_location_code='+this.selectorView.fixedRentalLocationValue();
+      }    
       if (commonServices.apiKey && commonServices.apiKey != '') {
         url += '&api_key='+commonServices.apiKey;
       }  
@@ -165,6 +206,11 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
       if (this.configuration.pickupReturnPlace && $(this.return_place_selector).val() != '') {
         url += '&place='+$(this.return_place_selector).val();
       }        
+      if (this.selectorView.applyRentalLocationSelector() && $(this.rental_location_code_selector).val() != '') {
+        url += '&rental_location_code='+$(this.rental_location_code_selector).val();
+      } else if (this.selectorView.usedFixedRentalLocation() && this.selectorView.fixedRentalLocationValue() != '') {
+        url += '&rental_location_code='+this.selectorView.fixedRentalLocationValue();
+      }   
       if (commonServices.apiKey && commonServices.apiKey != '') {
         url += '&api_key='+commonServices.apiKey;
       }        
@@ -193,6 +239,11 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
       if (this.configuration.pickupReturnPlace && $(this.pickup_place_selector).val() != '') {
         url += '&place='+$(this.pickup_place_selector).val();
       }         
+      if (this.selectorView.applyRentalLocationSelector() && $(this.rental_location_code_selector).val() != '') {
+        url += '&rental_location_code='+$(this.rental_location_code_selector).val();
+      } else if (this.selectorView.usedFixedRentalLocation() && this.selectorView.fixedRentalLocationValue() != '') {
+        url += '&rental_location_code='+this.selectorView.fixedRentalLocationValue();
+      }       
       if (commonServices.apiKey && commonServices.apiKey != '') {
         url += '&api_key='+commonServices.apiKey;
       }        
@@ -221,6 +272,11 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
       if (this.configuration.pickupReturnPlace && $(this.return_place_selector).val() != '') {
         url += '&place='+$(this.return_place_selector).val();
       }    
+      if (this.selectorView.applyRentalLocationSelector() && $(this.rental_location_code_selector).val() != '') {
+        url += '&rental_location_code='+$(this.rental_location_code_selector).val();
+      } else if (this.selectorView.usedFixedRentalLocation() && this.selectorView.fixedRentalLocationValue() != '') {
+        url += '&rental_location_code='+this.selectorView.fixedRentalLocationValue();
+      }   
       if (commonServices.apiKey && commonServices.apiKey != '') {
         url += '&api_key='+commonServices.apiKey;
       }       
@@ -281,6 +337,42 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
           }          
         }
       }
+
+    }
+
+    /**
+     * Rental location changed
+     */ 
+    this.rentalLocationChanged = function() {
+
+       console.log('rental location changed');
+
+       // Clear shopping cart data to avoid to be reloaded
+       if (this.selectorModel.shopping_cart) {
+         this.selectorModel.shopping_cart.date_from = null;
+         this.selectorModel.shopping_cart.time_from = null;
+         this.selectorModel.shopping_cart.date_to = null;
+         this.selectorModel.shopping_cart.time_to = null;
+       }  
+
+       // -- Disable fields selectors (return place, time from ,date to, time to)
+
+       // Disable time from
+       $(this.selectorModel.time_from_selector).attr('disabled', true);
+       // Disable date to
+       $(this.selectorModel.date_to_selector).attr('disabled', true);
+       // Disable time to
+       $(this.selectorModel.time_to_selector).attr('disabled', true);
+
+       // Initialize date from, time from, date to and time to
+       $(this.selectorModel.date_from_selector).datepicker('setDate', null);
+       if (this.selectorModel.configuration.timeToFrom) {
+         $(this.selectorModel.time_from_selector).val('');
+       }
+       $(this.selectorModel.date_to_selector).datepicker('setDate', null);
+       if (this.selectorModel.configuration.timeToFrom) {
+         $(this.selectorModel.time_to_selector).val('');
+       }
 
     }
 
@@ -393,6 +485,7 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
           else { // User selects pickup place from the list
               $(this.selectorModel.custom_return_place_selector).val('false');
               $(this.selectorModel.another_return_place_group_selector).hide();
+              $(this.selectorModel.return_place_group_selector).show();
           }
         }
 
@@ -511,29 +604,6 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
                         // Localize UI
                         //$('.nav').localize();
                      });
-        // Setup Form
-        this.setupSelectorFormTmpl();
-
-        // Setup UI
-        $(this.selectorModel.form_selector).attr('action', commonServices.chooseProductUrl);
-        
-        this.extractAgentId();
-
-        // Setup pickup/return places
-        if (this.selectorModel.configuration.pickupReturnPlace) {
-          this.setupPickupReturnPlace();
-        }
-        
-        // Setup dates
-        this.setupDateControls();
-        
-        // Setup time from / to
-        if (this.selectorModel.configuration.timeToFrom) {
-          this.setupTimeToFrom();
-        }       
-
-        // Setup validation
-        this.setupValidation();
 
         // Setup on scroll => Google Location API integration
         var self = this;
@@ -544,12 +614,38 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
 
     }
 
+    /**
+     *  Used fixed rental location
+     */ 
+    this.usedFixedRentalLocation = function() {
+      return $(this.selectorModel.form_selector).find('input[type=hidden][name=rental_location_code]').length > 0;
+    }
+
+    /**
+     *  Fixed rental location value
+     */ 
+    this.fixedRentalLocationValue = function() {
+      return $(this.selectorModel.form_selector).find('input[type=hidden][name=rental_location_code]').val();
+    }
+
+    /**
+     *  Check if rental location is not hidden
+     */ 
+    this.applyRentalLocationSelector = function() {
+
+      var not_hidden_rental_location_code = ($(this.selectorModel.form_selector).find('input[type=hidden][name=rental_location_code]').length == 0);
+      return (not_hidden_rental_location_code && this.selectorModel.configuration.selectRentalLocation);
+
+    }
+
     // ------------------------ Start -----------------------------------------
 
     /**
      * Start the component empty
      */
     this.startEmpty = function() {
+
+      this.prepareSelector();
 
       // Disable selector controls  
       $(this.selectorModel.pickup_place_selector).attr('disabled', true);
@@ -564,10 +660,16 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
         this.loadPickupPlaces();
       }
       else {
-        var date = moment();
-        this.selectorModel.loadPickupDays(date.year(), date.month()+1);
+        if (this.applyRentalLocationSelector()) {
+          this.loadRentalLocations();
+        }
+        else {
+          var date = moment();
+          this.selectorModel.loadPickupDays(date.year(), date.month()+1);
+        }
       }
 
+      // Load families
       if (this.selectorModel.configuration.selectFamily) {
         this.loadFamilies();
       }
@@ -580,6 +682,8 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
     this.startFromShoppingCart = function(shopping_cart) { /* Show the selector with the shopping cart information */
 
       this.selectorModel.shopping_cart = shopping_cart;
+
+      this.prepareSelector();
 
       // Initialize number of products
       $(this.selectorModel.number_of_products_selector).val(shopping_cart.number_of_products);
@@ -597,6 +701,12 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
         this.loadPickupPlaces(); // The other fields are automatically assigned after pickup_place assignation
       }
       else {
+
+        // Load rental locations
+        if (this.applyRentalLocationSelector()) {
+          this.loadRentalLocations();
+        }
+
         // No delivery/collection place => Directly assign date_from and date_to from shopping_cart
         var date_from = moment(shopping_cart.date_from).format(this.selectorModel.configuration.dateFormat); 
         var date_to = moment(shopping_cart.date_to).format(this.selectorModel.configuration.dateFormat); 
@@ -613,6 +723,42 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
       }
 
     }
+
+   /**
+     * Prepare the selector
+     */
+    this.prepareSelector = function() {
+
+        // Setup Form
+        this.setupSelectorFormTmpl();
+
+        // Setup UI
+        $(this.selectorModel.form_selector).attr('action', commonServices.chooseProductUrl);
+        
+        this.extractAgentId();
+
+        // Setup pickup/return places
+        if (this.selectorModel.configuration.pickupReturnPlace) {
+          this.setupPickupReturnPlace();
+        }
+        else {
+          if (this.applyRentalLocationSelector()) {
+            this.setupRentalLocations();
+          }         
+        }
+        
+        // Setup dates
+        this.setupDateControls();
+        
+        // Setup time from / to
+        if (this.selectorModel.configuration.timeToFrom) {
+          this.setupTimeToFrom();
+        }       
+
+        // Setup validation
+        this.setupValidation();
+    }
+
 
     // ------------------------ Extract Agent Id ------------------------------
 
@@ -660,14 +806,43 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
 
       // The selector form fields are defined in a micro-template
       if (document.getElementById(this.selectorModel.form_selector_tmpl)) {
-        // Check if forced hidden family_id
-        var not_hidden_family_id = ($(this.selectorModel.form_selector).find('input[type=hidden][name=family_id]').length == 0);
+        console.log(this.selectorModel.configuration);
+        var not_hidden_rental_location_code = false;
+        var not_hidden_family_id = false;
+        if (this.selectorModel.shopping_cart) {
+          not_hidden_family_id = !this.selectorModel.shopping_cart.engine_fixed_family;
+          not_hidden_rental_location_code = !this.selectorModel.shopping_cart.engine_fixed_rental_location;
+        }
+        else {
+          // Check if forced hidden rental_location_code
+          not_hidden_rental_location_code = ($(this.selectorModel.form_selector).find('input[type=hidden][name=rental_location_code]').length == 0);
+          // Check if forced hidden family_id
+          not_hidden_family_id = ($(this.selectorModel.form_selector).find('input[type=hidden][name=family_id]').length == 0);
+        }
+
         // Load the template
         var html = tmpl(this.selectorModel.form_selector_tmpl)({configuration: this.selectorModel.configuration,
-                                                                not_hidden_family_id: not_hidden_family_id });
+                                                                not_hidden_family_id: not_hidden_family_id,
+                                                                not_hidden_rental_location_code: not_hidden_rental_location_code,
+                                                                rental_location_code: (this.selectorModel.shopping_cart ? this.selectorModel.shopping_cart.rental_location_code : null),
+                                                                family_id:  (this.selectorModel.shopping_cart ? this.selectorModel.shopping_cart.family_id : null) });
         // Assign to the form
         $(this.selectorModel.form_selector).append(html);
       }
+
+    }
+
+    /**
+     *  Setup rental location control
+     */ 
+    this.setupRentalLocations = function() {
+
+      var self = this;
+
+      // Bind pickup place changed
+      $(this.selectorModel.rental_location_code_selector).bind('change', function() {
+         self.selectorController.rentalLocationChanged();
+      });
 
     }
 
@@ -950,6 +1125,9 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
              $(self.selectorModel.form_selector + ' label.form-reservation-error').remove();
            },
            rules: {
+               rental_location_code: {
+                   required: self.selectorModel.rental_location_code_selector + ':visible'
+               },
                pickup_place: {
                    required: self.selectorModel.pickup_place_selector + ':visible'
                },
@@ -997,6 +1175,9 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
                }               
            },
            messages: {
+               rental_location_code: {
+                   required: i18next.t('common.required')
+               },
                pickup_place: {
                    required: i18next.t('selector.validations.pickupPlaceRequired')
                },
@@ -1059,6 +1240,15 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
     }
 
     // ------------------------ Load data -------------------------------------
+
+    /**
+     *  Setup rental locations
+     */ 
+    this.loadRentalLocations = function() {
+
+      this.selectorModel.loadRentalLocations();
+
+    }
 
     /**
      * Setup the families
@@ -1294,7 +1484,8 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
                       }
                       else {
                         $(self.selectorModel.return_place_selector).val($(self.selectorModel.pickup_place_selector).val());
-                        $(self.selectorModel.return_place_selector).trigger('change');
+                        $(self.selectorModel.return_place_other_selector).val('');
+                        //$(self.selectorModel.return_place_selector).trigger('change');
                       }
                       // In both cases notify that the return place has changed
                       self.selectorController.returnPlaceChanged();
@@ -1333,6 +1524,21 @@ define('SelectorRent', ['jquery', 'YSDMemoryDataSource', 'YSDRemoteDataSource','
             $(this.selectorModel.family_selector).show();
           }
           break;
+        case 'rentalLocations':
+          if ($(this.selectorModel.rental_location_code_selector).length > 0 && this.selectorModel.rentalLocations.length > 0) {
+            var dataSource = new MemoryDataSource(this.selectorModel.rentalLocations);
+            var rental_location_code = this.selectorModel.shopping_cart ? this.selectorModel.shopping_cart.rental_location_code : null;
+            console.log('rental_location_code');
+            console.log(this.selectorModel.shopping_cart);
+            console.log(rental_location_code);
+            var rentalLocationCode = new SelectSelector(this.selectorModel.rental_location_code,
+                                                        dataSource, 
+                                                        rental_location_code, 
+                                                        true,
+                                                        i18next.t('selector.select'));
+            $(this.selectorModel.rental_location_selector).show();
+          }          
+          break;  
         case 'place':
           if (id== 'pickup_place') {
             // Enable return place selector